Fix YAML imports and admin page rendering
