+++ app/api/cms-config/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import YAML from "yaml";
 
-import configTemplate from "@/admin/config.yml";
+import { loadConfigTemplate } from "@/lib/cms/config-template";
 
 export const dynamic = "force-dynamic";
 export const revalidate = 0;
@@ -15,7 +15,7 @@ function isRecord(value: unknown): value is Record<string, unknown> {
 }
 
 function parseConfig(): ConfigData {
-  const parsed = YAML.parse(configTemplate);
+  const parsed = YAML.parse(loadConfigTemplate());
 
   if (!parsed || typeof parsed !== "object") {
     throw new Error("Invalid CMS configuration template.");
