+++ lib/cms/config-template.ts
@@ -1,0 +1,5 @@
+import { loadStaticFile } from "@/lib/server/static-file";
+
+export function loadConfigTemplate(): string {
+  return loadStaticFile("admin/config.yml");
+}
